Testing Poll SCM

+++ src/main/java/com/example/demo/CalculatorController.java
@@ -1,14 +1,17 @@
 package com.example.demo;
 
 
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/calc")
 public class CalculatorController {
 
     @GetMapping("/add")
-    public int add(@RequestParam int a, @RequestParam int b) {
+    public double add(@RequestParam double a, @RequestParam double b) {
         return a + b;
     }
